+++ src/lib/pages/tx-details/components/MessageSection.tsx
@@ -34,7 +34,7 @@ export const MessageSection = ({ txData }: MessageSectionProps) => {
               boxSize={4}
             />
             <AlertDescription wordBreak="break-word">
-              {txData.raw_log}
+              {txData.rawLog}
             </AlertDescription>
           </Flex>
         </Alert>
@@ -43,11 +43,11 @@ export const MessageSection = ({ txData }: MessageSectionProps) => {
         <Heading as="h6" variant="h6">
           Messages
         </Heading>
-        <Badge>{messages.length}</Badge>
+        <Badge>{messages?.length}</Badge>
       </Flex>
-      {messages.map((msg, idx) => (
+      {messages?.map((msg, idx) => (
         <TxMessage
-          key={msg.type + msg.value + idx.toString()}
+          key={JSON.stringify(msg) + idx.toString()}
           msgBody={msg}
           log={logs[idx]}
           isSingleMsg={messages.length === 1}