+++ src/lib/services/types/index.ts
@@ -2,6 +2,7 @@ export * from "./bank";
 export * from "./block";
 export * from "./distribution";
 export * from "./proposal";
+export * from "./protobuf";
 export * from "./staking";
 export * from "./validator";
 export * from "./wasm";